+++ showcase/harness/README.md
@@ -247,7 +247,7 @@ Every `kind` resolves to a driver registered in `src/probes/drivers/index.ts`. T
 
 | YAML `kind`             | Driver file                        | Emitted key prefix(es)                 | Shape     |
 | ----------------------- | ---------------------------------- | -------------------------------------- | --------- |
-| `smoke`                 | `drivers/liveness.ts`              | `smoke:<slug>` **and** `health:<slug>` | static    |
+| `smoke`                 | `drivers/d2-liveness.ts`           | `smoke:<slug>` **and** `health:<slug>` | static    |
 | `e2e_smoke` (deferred)  | `drivers/e2e-chat-tools.ts`        | `e2e_smoke:<suite>`                    | static    |
 | `image_drift`           | `drivers/image-drift.ts`           | `image_drift:<service>`                | discovery |
 | `version_drift`         | `drivers/version-drift.ts`         | `version_drift:<pkg>`                  | discovery |
