+++ showcase/harness/config/probes/e2e-demos.yml
@@ -7,7 +7,7 @@
 # declared in the service's registry entry. One chromium per tick fans
 # out serially across all demos of all in-scope services.
 #
-# Rows emitted per driver invocation (see e2e-readiness.ts):
+# Rows emitted per driver invocation (see d3-readiness.ts):
 #   - Primary `e2e-demos:<slug>`           ProbeResult carrying the
 #                                          aggregate { total, passed,
 #                                          failed[], shape, backendUrl }
@@ -41,19 +41,19 @@
 # ── timeout_ms: 1_200_000 (20 min) ─────────────────────────────────────
 #
 # Outer driver-invocation cap. The driver subdivides internally with a
-# 30s per-page timeout (DEFAULT_PAGE_TIMEOUT_MS in e2e-readiness.ts) — the
+# 30s per-page timeout (DEFAULT_PAGE_TIMEOUT_MS in d3-readiness.ts) — the
 # largest 38-demo service could exceed 15 min on cold starts, so the
 # outer cap is set to 20 min. The driver's internal hard-cap is
 # threaded from this YAML value at orchestrator boot (see
 # orchestrator.ts: `process.env.E2E_DEMOS_TIMEOUT_MS = cfg.timeout_ms`,
-# read per-`run()` by drivers/e2e-readiness.ts), so the driver's own timer
+# read per-`run()` by drivers/d3-readiness.ts), so the driver's own timer
 # fires at exactly the same wall-clock as the invoker's outer race —
 # no premature internal-cap edge that would short-circuit the YAML's
 # documented behaviour at the driver's pre-threading default of 5 min.
 # When the cap fires inside the driver, the driver itself iterates the
 # remaining demos and side-emits `errorClass: "abort"` /
 # `errorDesc: "timeout after 1200000ms"` rows for each — the per-demo
-# fan-out loop in `e2e-readiness.ts` checks `abort.signal.aborted` at the
+# fan-out loop in `d3-readiness.ts` checks `abort.signal.aborted` at the
 # top of every iteration and writes a red side-row before continuing
 # to the next demo. The dashboard keeps rendering red dots rather
 # than going gray, which is the right signal for operator triage
